Move schema and embed generation
