+++ util/smash/schema/event.gql
