+++ util/smash/schema/tourney.gql
