+++ util/smash/schema/tourneyState.gql
